TERRABIA-MOBILE/terrabia-backend/requirements.txt
# Core Django
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1

# Database (POSTGRESQL pour Railway)
psycopg2-binary==2.9.9

# Server
gunicorn==21.2.0

# Static files
whitenoise==6.6.0

# CORS
django-cors-headers==4.9.0

# Environment (si vous l'utilisez)
python-decouple==3.8

# WebSockets (UNIQUEMENT si vous en avez vraiment besoin)
# channels==4.3.2
# channels-redis==4.3.0
# daphne==4.2.1
# redis==7.1.0

# Optionnel
django-filter==25.2
drf-yasg==1.21.11
pillow==12.0.0

whitenoise==6.6.0
dj-database-url==2.1.0
